streamline testing with multiple forks
This slightly simplifies the gradle build for testing in the presence of
multiple forks, and also simplifies `CorsetValidator` as well.
