+++ arithmetization/src/main/java/net/consensys/linea/corset/CorsetValidator.java
@@ -15,118 +15,293 @@
 
 package net.consensys.linea.corset;
 
+import static net.consensys.linea.zktracer.ChainConfig.MAINNET_LONDON_TESTCONFIG;
+import static net.consensys.linea.zktracer.Trace.*;
+
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectWriter;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import net.consensys.linea.zktracer.ChainConfig;
+import net.consensys.linea.zktracer.Fork;
 
 /**
- * Responsible for running the command-line <code>corset</code> tool to check that a given trace is
- * accepted by the zkevm constraints. The <code>corset</code> tool has variables levels of
- * "expansion" which it can apply before performing the check. Greater levels of expansion imply
- * more accurate checks (i.e. more realistic compared to the prover). Furthermore, <code>corset
- * </code> can be configured to use field arithmetic or simply big integers (with the latter option
- * intended to offer faster but much less precise checking). The default configuration is set to
- * give good accuracy, but without significant overhead. A greater level can be configured by
- * enabling the <code>autoConstraints</code>.
+ * Responsible for running the command-line <code>go-corset</code> tool to check that a given trace
+ * is accepted by the zkevm constraints. The <code>corset</code> tool has three IR levels at which
+ * it can do this (HIR/MIR/AIR). The lowest level (AIR) provides the most accurate results (i.e.
+ * most comparable with prover).
  *
- * <p>The configuration can be set using the environment variable <code>CORSET_FLAGS</code>. Example
- * values for this environment variable include <code>fields,expand</code> (sets <code>
- * fieldArithmetic=true</code> and <code>expansion=0</code>) and <code>fields,expand,expand,
- * auto</code> (enables <code>fieldArithmetic</code> and <code>autoConstraints</code> and sets
- * <code>expansion=2</code>). Note, it doesn't make sense to have <code>expand</code> without <code>
- * fields</code>. Likewise, it doesn't make sense to have <code>auto</code> without <code>expand
- * </code>.
+ * <p>The configuration can be set using the environment variable <code>GO_CORSET_FLAGS</code>. If
+ * this is not set, then this validator will not be activated. </code>.
  */
 @Slf4j
-public class CorsetValidator {
+public class CorsetValidator extends AbstractExecutable {
   public record Result(boolean isValid, File traceFile, String corsetOutput) {}
 
-  /** */
-  private static final String ZK_EVM_RELATIVE_PATH = "/linea-constraints/";
-
-  // TODO: make it dynamic
-
-  private static final String ZK_EVM_BIN = "zkevm_london.bin";
-
-  /** Specifies the default zkEVM.bin file to use (including its path). */
-  private static String defaultZkEvm = null;
-
-  /** Interface to Go corset tool. */
-  private final GoCorsetValidator goCorset;
+  private static final String ZKEVM_BIN_PREFIX = "../linea-constraints/";
+  private static final String[] ZKEVM_BINS;
 
   static {
-    // Configure default path to the zkevm.bin file.
-    initDefaultZkEvm();
+    ZKEVM_BINS = new String[Fork.values().length];
+    ZKEVM_BINS[Fork.LONDON.ordinal()] = "zkevm_london.bin";
+    ZKEVM_BINS[Fork.PARIS.ordinal()] = "zkevm_paris.bin";
+    ZKEVM_BINS[Fork.SHANGHAI.ordinal()] = "zkevm_shanghai.bin";
+    ZKEVM_BINS[Fork.CANCUN.ordinal()] = "zkevm_cancun.bin";
+    ZKEVM_BINS[Fork.PRAGUE.ordinal()] = "zkevm_prague.bin";
   }
+
+  /** Indicates whether or not this validator is active (i.e. we located the go-corset binary). */
+  @Getter private boolean active = false;
+
+  /** Details chain configuration. */
+  private final ChainConfig chain;
+
+  /**
+   * Aggregation of all coverage reports generated by go-corset. If GOCORSET_COVERAGE is set, then
+   * this data will be written out to that file upon JVM termination
+   */
+  private static final CoverageReport coverageReport = new CoverageReport();
 
   public CorsetValidator(ChainConfig chain) {
-    // Construct and initialise Go corset.
-    this.goCorset = new GoCorsetValidator(chain);
+    this.chain = chain;
+    configGoCorset();
   }
 
   /**
-   * Attempt to validate a given tracefile against a given set of zkEVM constraints. A default
-   * location for the zkevm.bin file is used.
+   * Attempt to validate a given trace against a given zkEvmBin file.
    *
-   * @param traceFile The tracefile being validated.
+   * @param traceFile Path to the trace file being validated.
    * @return A result which tells us whether or not the trace file was accepted, and provides
    *     additional information for debugging purposes.
    */
-  public Result validate(final Path traceFile) throws RuntimeException {
-    if (defaultZkEvm == null) {
-      throw new IllegalArgumentException("Default zkevm.bin not set.");
+  public Result validate(final Path traceFile) {
+    String zkevmBin = ZKEVM_BIN_PREFIX + ZKEVM_BINS[chain.fork.ordinal()];
+    //
+    if (active) {
+      Outcome outcome;
+      try {
+        List<String> commands = buildCommandLine(traceFile, zkevmBin);
+        log.debug("{}", commands);
+        // Execute corset with a 5s timeout.
+        outcome = super.exec(5, commands);
+      } catch (Throwable e) {
+        log.error("Corset validation has thrown an exception: %s".formatted(e.getMessage()));
+        throw new RuntimeException(e);
+      }
+      // Manage coverage report
+      consumeCoverageReport(traceFile);
+      // Check for success or failure
+      if (outcome.exitcode() != 0) {
+        log.error("Validation failed: %s".formatted(outcome.output()));
+        return new Result(false, traceFile.toFile(), outcome.output());
+      }
+      // success!
+      return new Result(true, traceFile.toFile(), outcome.output());
     }
-    return validate(traceFile, defaultZkEvm);
+    // Tool is not active
+    throw new RuntimeException("go-corset inactive!");
   }
 
   /**
-   * Attempt to validate a given tracefile against a given set of zkEVM constraints.
-   *
-   * @param traceFile The tracefile being validated.
-   * @param zkEvmBin The zkEVM constraints file (compiled using corset).
-   * @return A result which tells us whether or not the trace file was accepted, and provides
-   *     additional information for debugging purposes.
+   * Configure corset from the <code>CORSET_FLAGS</code> environment variable (if set). If the
+   * environment variable is not set, the default configuration is retained. If the environment
+   * variable is set, but its value is malformed then an exception is raised.
    */
-  public Result validate(final Path traceFile, final String zkEvmBin) {
-    // Generate result from Go corset tool.
-    Result rg = goCorset.validate(traceFile, zkEvmBin);
-    // Sanity check at least one validator is active
-    if (!goCorset.isActive()) {
-      throw new RuntimeException("go-corset not available");
-    } else {
-      return rg;
+  private void configGoCorset() {
+    // If we can execute go-corset then use it!
+    this.active = super.isExecutable("go-corset", "--help");
+  }
+
+  /**
+   * Construct the command-line to use for running corset to check a given trace file is accepted
+   * (or not) by a given bin file. Amongst other things, this will configure the command-line flags
+   * as dictated by the CORSET_FLAGS environment variable.
+   *
+   * @return
+   */
+  private List<String> buildCommandLine(Path traceFile, String zkEvmBin) {
+    ArrayList<String> options = new ArrayList<>();
+    // Determine options to use (either default or override)
+    String flags =
+        System.getenv().getOrDefault("GOCORSET_FLAGS", "--report --report-context 2 --hir ");
+    // Determine whether to generate a coverage report (or not).
+    String coverage = System.getenv().get("GOCORSET_COVERAGE");
+    // Specify corset binary
+    options.add("go-corset");
+    // Set chain properties
+    setChainConstants(options);
+    // Specify corset "check" command.
+    options.add("check");
+    // Add all options
+    options.addAll(Arrays.asList(flags.split(" ")));
+    // Add coverage flag (if applicable)
+    if (coverage != null) {
+      Path coverageFilename = determineCoverageFile(traceFile);
+      options.add("--coverage");
+      options.add(coverageFilename.toString());
+    }
+    // Specify trace file to use
+    options.add(traceFile.toAbsolutePath().toString());
+    // Specify the zkevm.bin file.
+    options.add(zkEvmBin);
+    // Done
+    return options;
+  }
+
+  /**
+   * Configure the constants as appropriate for the given chain. For Linea mainnet, do nothing as
+   * the defaults should match (and are what is used by the prover).
+   *
+   * @param options
+   */
+  private void setChainConstants(List<String> options) {
+    final int fork =
+        switch (chain.fork) {
+          case LONDON -> EVM_LONDON;
+            // case PARIS -> EVM_PARIS;
+          case SHANGHAI -> EVM_SHANGHAI;
+          case CANCUN -> EVM_CANCUN;
+          case PRAGUE -> EVM_PRAGUE;
+          default -> throw new IllegalArgumentException("unknown fork \"" + chain.fork + "\"");
+        };
+    options.add("-SEVM_FORK=" + fork);
+    if (chain != MAINNET_LONDON_TESTCONFIG) {
+      options.add("-Sblockdata.GAS_LIMIT_MINIMUM=" + chain.gasLimitMinimum.toString());
+      options.add("-Sblockdata.GAS_LIMIT_MAXIMUM=" + chain.gasLimitMaximum.toString());
+      if (chain.fixedGasLimitEnabled) {
+        options.add("-Sblockdata.GAS_LIMIT_ENABLE=1");
+      } else {
+        options.add("-Sblockdata.GAS_LIMIT_ENABLE=0");
+      }
     }
   }
 
-  private static void initDefaultZkEvm() {
-    final String currentDir;
+  /**
+   * Determine the name of a temporary file to use for writing the coverage report. For now, this is
+   * based upon the name of the trace file (but it doesn't have to be done like this).
+   *
+   * @param traceFile
+   * @return
+   */
+  private Path determineCoverageFile(Path traceFile) {
+    String filename = traceFile.toString().replace(".lt", ".coverage.json");
+    return Paths.get(filename);
+  }
 
+  private void consumeCoverageReport(Path traceFile) {
+    boolean coverage = System.getenv("GOCORSET_COVERAGE") != null;
     try {
-      currentDir = Path.of(".").toRealPath().toString();
-    } catch (final IOException e) {
-      log.error("Error while getting current directory: %s".formatted(e.getMessage()));
-      throw new RuntimeException(e);
+      if (coverage) {
+        Path coverageFile = determineCoverageFile(traceFile);
+        // Read in the coverage file
+        byte[] bytes = Files.readAllBytes(coverageFile);
+        // Needed for Jackson
+        ObjectMapper mapper = new ObjectMapper();
+        // Read bytes using Jackson
+        JsonNode json = mapper.readTree(bytes);
+        // Aggregate coverage data together
+        coverageReport.importJsonReport(json);
+        // Delete temporary coverage file
+        Files.delete(coverageFile);
+        //
+        log.debug("read coverage from {} ({}KB)", coverageFile, bytes.length / 1024);
+      }
+    } catch (IOException e) {
+      log.debug("error reading coverage report: " + e.getMessage());
     }
-
-    final String zkEvmBinInCurrentDir = currentDir + ZK_EVM_RELATIVE_PATH + binName();
-    if (new File(zkEvmBinInCurrentDir).exists()) {
-      defaultZkEvm = zkEvmBinInCurrentDir;
-      return;
-    }
-
-    final String zkEvmBinInDirAbove = currentDir + "/.." + ZK_EVM_RELATIVE_PATH + binName();
-    if (new File(zkEvmBinInDirAbove).exists()) {
-      defaultZkEvm = zkEvmBinInDirAbove;
-      return;
-    }
-
-    log.warn("Could not find default path for {}", binName());
   }
 
-  private static String binName() {
-    return System.getenv("ZKEVM_BIN") != null ? System.getenv("ZKEVM_BIN") : ZK_EVM_BIN;
+  static {
+    String coverageFile = System.getenv("GOCORSET_COVERAGE");
+    //
+    if (coverageFile != null) {
+      // Coverage is being requested, therefore register a shutdown hook so that we can aggregate
+      // all the reports at the end.
+      Thread aggregationHook =
+          new Thread(
+              () -> {
+                aggregateCoverageReports(Paths.get(coverageFile));
+              });
+      Runtime.getRuntime().addShutdownHook(aggregationHook);
+    }
+  }
+
+  static void aggregateCoverageReports(Path coverageFile) {
+    try {
+      byte[] bytes = coverageReport.toJsonBytes();
+      // Finally, write the aggregated report out to the destination file.
+      Files.write(coverageFile, bytes);
+      // Record what happened
+      log.debug("wrote aggregated coverage report to {} ({}KB)", coverageFile, bytes.length / 1024);
+    } catch (IOException e) {
+      log.error("error writing aggregated coverage report: {}", e.getMessage());
+    }
+  }
+
+  static class CoverageReport {
+    HashMap<String, HashSet<Integer>> air = new HashMap<>();
+    HashMap<String, HashSet<Integer>> mir = new HashMap<>();
+    HashMap<String, HashSet<Integer>> hir = new HashMap<>();
+
+    public synchronized void importJsonReport(JsonNode report) throws IOException {
+      importJsonSection(report.get("air"), air);
+      importJsonSection(report.get("mir"), mir);
+      importJsonSection(report.get("hir"), hir);
+    }
+
+    public synchronized byte[] toJsonBytes() throws IOException {
+      var report = new HashMap<>();
+      if (!air.isEmpty()) {
+        report.put("air", air);
+      }
+      if (!mir.isEmpty()) {
+        report.put("mir", mir);
+      }
+      if (!hir.isEmpty()) {
+        report.put("hir", hir);
+      }
+      ObjectWriter mapper = new ObjectMapper().writer().withDefaultPrettyPrinter();
+      return mapper.writeValueAsBytes(report);
+    }
+
+    private void importJsonSection(JsonNode section, Map<String, HashSet<Integer>> data) {
+      // section can be null if there was no recorded data for the given target ir.
+      if (section != null) {
+        // Iterate over the fields of the section
+        for (Iterator<Map.Entry<String, JsonNode>> it = section.fields(); it.hasNext(); ) {
+          Map.Entry<String, JsonNode> field = it.next();
+          HashSet<Integer> fieldData = data.get(field.getKey());
+          // Initialise field data if necessary
+          if (fieldData == null) {
+            fieldData = new HashSet<>();
+            data.put(field.getKey(), fieldData);
+          }
+          // Java happiness
+          final HashSet<Integer> nFieldData = fieldData;
+          // Add all elements to the set)
+          field
+              .getValue()
+              .elements()
+              .forEachRemaining(
+                  (JsonNode n) -> {
+                    nFieldData.add(n.asInt());
+                  });
+        }
+      }
+    }
   }
 }
