+++ gradle/tests.gradle
@@ -127,7 +127,7 @@ jacocoTestReport {
 }
 
 tasks.test.configure {
-  dependsOn(buildZkevmBin, "copyPlugins")
+  dependsOn("copyPlugins")
 
   delete("${project.getLayout().getBuildDirectory().get()}/besu/traces")
   mkdir("${project.getLayout().getBuildDirectory().get()}/besu/traces")
@@ -170,8 +170,7 @@ tasks.register("copyPlugins", Copy) {
 }
 
 tasks.register("fastReplayTests", Test) {
-  dependsOn(buildZkevmBin)
-    //
+  //
   if (System.getenv().containsKey("REPLAY_TESTS_PARALLELISM")) {
     systemProperty("junit.jupiter.execution.parallel.enabled", true)
     systemProperty("junit.jupiter.execution.parallel.mode.default", "concurrent")
@@ -198,8 +197,6 @@ tasks.register("jacocoFastReplayTestsReport", JacocoReport) {
 }
 
 tasks.register("nightlyReplayTests", Test) {
-  dependsOn(buildZkevmBin)
-
   if (System.getenv().containsKey("REPLAY_TESTS_PARALLELISM")) {
     systemProperty("junit.jupiter.execution.parallel.enabled", true)
     systemProperty("junit.jupiter.execution.parallel.mode.default", "concurrent")
@@ -216,8 +213,6 @@ tasks.register("nightlyReplayTests", Test) {
 }
 
 tasks.register("nightlyTests", Test) {
-  dependsOn(buildZkevmBin)
-
   systemProperty("junit.jupiter.execution.parallel.enabled", true)
   systemProperty("junit.jupiter.execution.parallel.mode.default", "concurrent")
   systemProperty("junit.jupiter.execution.parallel.mode.classes.default", "concurrent")
@@ -241,8 +236,6 @@ tasks.register("jacocoNightlyTestsReport", JacocoReport) {
 }
 
 tasks.register("weeklyTests", Test) {
-  dependsOn(buildZkevmBin)
-
   systemProperty("junit.jupiter.execution.parallel.enabled", true)
   systemProperty("junit.jupiter.execution.parallel.mode.default", "concurrent")
   systemProperty("junit.jupiter.execution.parallel.mode.classes.default", "concurrent")
@@ -266,8 +259,6 @@ tasks.register("jacocoWeeklyTestsReport", JacocoReport) {
 }
 
 tasks.register("prcCallTests", Test) {
-  dependsOn(buildZkevmBin)
-  //
   systemProperty("junit.jupiter.execution.parallel.enabled", true)
   systemProperty("junit.jupiter.execution.parallel.mode.default", "concurrent")
   systemProperty("junit.jupiter.execution.parallel.mode.classes.default", "concurrent")
