+++ reference-tests/build.gradle
@@ -57,7 +57,7 @@ tasks.register("generateGeneralStateReferenceTests", RefTestGenerationTask) {
 }
 
 tasks.register('referenceBlockchainTests', Test) {
-  dependsOn(":arithmetization:buildZkevmBin")
+  //dependsOn(":arithmetization:buildZkevmBin")
   finalizedBy(jacocoReferenceBlockchainTestsReport)
   description = 'Runs ETH reference blockchain tests.'
   dependsOn generateBlockchainReferenceTests
@@ -82,11 +82,10 @@ tasks.register("jacocoReferenceBlockchainTestsReport", JacocoReport) {
 }
 
 tasks.register('referenceGeneralStateTests', Test) {
-  dependsOn(":arithmetization:buildZkevmBin")
+  //dependsOn(":arithmetization:buildZkevmBin")
   finalizedBy(jacocoReferenceGeneralStateTestsReport)
   description = 'Runs ETH reference general state tests.'
   dependsOn generateGeneralStateReferenceTests
-  // Record zkevm.bin for CorsetValidator
   environment.put("REFERENCE_TEST_OUTCOME_OUTPUT_FILE", "GeneralStateReferenceTestOutcome.json")
 
   useJUnitPlatform {
